pyproject.toml
[project]
name = "trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ivan Osipchyk", email = "ivan.osipchyk.work@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.48",
    "ctrader-open-api>=0.9.2",
    "langchain-anthropic>=1.4.0",
    "langchain-openai>=0.3.0",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[project.scripts]
trading-validate = "trading.gui_validation:main"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pandas-stubs>=3.0.0.260204",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes (unused imports, undefined names)
    "I",   # isort (import ordering)
    "UP",  # pyupgrade (modernize syntax)
]

[tool.mypy]
python_version = "3.13"
strict = true
files = ["src"]

[[tool.mypy.overrides]]
module = "ccxt"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "ccxt",
    "langchain_anthropic",
    "langchain_core.*",
    "langchain_openai",
]
ignore_missing_imports = true
ignore_errors = true
